added next-question function and moved grid-templates
