--- code/src/Submit.js
@@ -1,6 +1,0 @@
-import React from 'react';
-
-export const Submit = () => {
-    return <button tabIndex={'0'} type="submit">Submit Survey</button>
-}
-
