+++ code/src/components/App.js
@@ -1,0 +1,7 @@
+import React from "react";
+import { Survey } from "./Survey";
+import "./app.css";
+
+export const App = () => {
+  return <Survey />
+};
