+++ code/src/components/app.css
@@ -9,6 +9,7 @@
 .app__wrapper {
     margin: 20px;
     padding: 20px;
+    
 }
 
 h2 {
