+++ code/src/components/text.css
@@ -1,0 +1,3 @@
+.text-input {
+    opacity: 0.8;
+}
