+++ code/src/index.css
@@ -5,13 +5,6 @@ body {
     sans-serif;
   -webkit-font-smoothing: antialiased;
   -moz-osx-font-smoothing: grayscale;
-  display: grid;
-    grid-template-rows: min-content;
-    grid-template-columns: min-content;
-    justify-content: center;
-    background-image: url('./pineapple-supply-co-prP9CYcRpzA-unsplash.jpg');
-    object-fit: contain;
-    background-position: center;
 }
 
 code {
