--- code/src/survey.css
@@ -1,7 +1,0 @@
-
-.survey__form {
-    grid-row: 2;
-    grid-column: 2/3;
-    background-color: rgba(247, 249, 251, 0.4);
-    padding: 20px;
-}
